Archive of CAP Data Available in July 2018
These three files were the latest CAP data available in July 2018. As discussed in the README, it is not possible to tell exactly when the new additions to these data were made. The .csv file for demographic data on complainants has been corrected to fix a mistake in the data provided online by the City of Philadelphia. The City accidentally switched the data for complainants ages and initials; we have fixed this mistake in our data.